First setup custom logging.
