+++ dotnet/Logging/ILogWriterFactory.cs
@@ -1,0 +1,7 @@
+﻿namespace WinSCP.Internal
+{
+    public interface ILogWriterFactory
+    {
+        ILogWriter Create(string fileName);
+    }
+}
