+++ dotnet/WinSCPnet.csproj
@@ -60,15 +60,20 @@
     <Compile Include="Internal\CallstackAndLock.cs" />
     <Compile Include="Internal\ConsoleCommStruct.cs" />
     <Compile Include="Internal\ExeSessionProcess.cs" />
+    <Compile Include="Logging\FileLogWriter.cs" />
+    <Compile Include="Logging\FileLogWriterFactory.cs" />
     <Compile Include="Internal\GenericSecurity.cs" />
+    <Compile Include="Logging\ILogWriter.cs" />
+    <Compile Include="Logging\ILogWriterFactory.cs" />
     <Compile Include="Internal\Job.cs" />
     <Compile Include="Internal\Lock.cs" />
-    <Compile Include="Internal\Logger.cs" />
+    <Compile Include="Logging\Logger.cs" />
     <Compile Include="Internal\PatientFileStream.cs" />
     <Compile Include="Internal\ProgressHandler.cs" />
     <Compile Include="Internal\SessionElementLogReader.cs" />
     <Compile Include="Internal\Tools.cs" />
     <Compile Include="Internal\UnsafeNativeMethods.cs" />
+    <Compile Include="LogWriterFactory.cs" />
     <Compile Include="OutputDataReceivedEventArgs.cs" />
     <Compile Include="FileTransferProgressEventArgs.cs" />
     <Compile Include="QueryReceivedEventArgs.cs" />
@@ -92,8 +97,8 @@
     <Compile Include="TransferEventArgs.cs" />
     <Compile Include="TransferOperationResult.cs" />
     <Compile Include="TransferOptions.cs" />
-    <Compile Include="Internal\CustomLogReader.cs" />
-    <Compile Include="Internal\ElementLogReader.cs" />
+    <Compile Include="Logging\CustomLogReader.cs" />
+    <Compile Include="Logging\ElementLogReader.cs" />
     <Compile Include="Internal\OperationResultGuard.cs" />
     <Compile Include="OperationEventArgs.cs" />
     <Compile Include="OperationResultBase.cs" />
